WIP: refactor named instances
This changes the flow of information:

- moku node has named_instance resources that:
  - creates concat fragment resources for the moku config on the moku node
  - export 'named_instance::app' and 'named_instance::proxy' resources

- nodes that include profile::named_instances:
  - import the named_instance::app resource

This both:

- ensures there is a clear responsibility for the flow of
information and that each piece of information has a single resource
responsible for it

- eliminates the use of ensure_resources (so long as there is a single
moku node responsible for deploying any particular app, anyway)

At the named_instance level, some of the parameters used only by the app
or proxy are bundled into app and proxy hashes, which helps reduce the
number of parameters in named_instance and clarify the purpose of some
of the parameters for named instances in hiera.
